Move position calculation from DashboardView to usePositionStats hook
